fix: implement input validation checks & modularize add and search patient functions
Accounted for cases where user enters a String when scanf is expecting an integer - one call of getChar() is not sufficient to clear the input buffer in this case, so I added clearInputBuffer(). Currently searchForPatientRecord() does not have proper validation checks for the menu as I'm calling getSelection() from menu.c, which does not have access to clearInputBuffer(). I'm too scared to mess with the files at this point in time so this will be addressed in a future version.